chore: init v2 routes
