+++ packages/chatsift/api-wrapper/src/v2/auth/ghost.ts
@@ -1,0 +1,7 @@
+// /api/v2/auth/ghost
+
+// GET /
+export type GetAuthGhostResult = never;
+
+// GET /callback
+export type GetAuthGhostCallbackResult = never;
