+++ packages/chatsift/api-wrapper/src/v2/index.ts
@@ -1,2 +1,3 @@
 export * from '../common';
 export * from './auth/discord';
+export * from './auth/ghost';
