+++ packages/chatsift/api-wrapper/v1.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line @typescript-eslint/no-require-imports
+module.exports = require('./dist/v1');
