+++ services/api/package.json
@@ -41,6 +41,7 @@
 		"@automoderator/filter-ignores": "workspace:^0.1.0",
 		"@automoderator/injection": "workspace:^0.1.0",
 		"@automoderator/logger": "workspace:^0.1.0",
+		"@chatsift/api-wrapper": "workspace:^0.1.0",
 		"@chatsift/readdir": "workspace:^0.1.0",
 		"@chatsift/rest-utils": "workspace:^0.1.0",
 		"@cordis/bitfield": "^1.2.0",
