+++ services/api/src/index.ts
@@ -16,12 +16,12 @@ import helmet from 'helmet';
 import { logRequests } from './middleware';
 import { TokenManager } from './util';
 import { PrismaClient } from '@prisma/client';
-const prisma = new PrismaClient();
 
 void (async () => {
 	const config = initConfig();
 	const logger = createLogger('api');
 
+	container.register(PrismaClient, { useValue: new PrismaClient() });
 	container.register(kLogger, { useValue: logger });
 	container.register(kSql, {
 		useValue: postgres(config.dbUrl, {
