+++ services/api/src/routes/api/v1/auth/discord/refresh/get.ts
@@ -1,52 +1,3 @@
-import { discordOAuth2 } from '#util';
-import type { AuthGetDiscordRefreshBody } from '@automoderator/core';
-import { Config, kConfig } from '@automoderator/injection';
-import { unauthorized } from '@hapi/boom';
-import cookie from 'cookie';
-import * as zod from 'zod';
-import type { NextHandler, Request, Response } from 'polka';
-import { inject, injectable } from 'tsyringe';
-import { jsonParser, Route, validate } from '@chatsift/rest-utils';
+// API v1 is now deprecated. This route is unchanged, and as such the same class is re-exported from v2.
 
-@injectable()
-export default class extends Route {
-	public override readonly middleware = [
-		jsonParser(),
-		validate(
-			zod
-				.object({
-					refresh_token: zod.string(),
-				})
-				.optional(),
-			'body',
-		),
-	];
-
-	public constructor(@inject(kConfig) public readonly config: Config) {
-		super();
-	}
-
-	public async handle(req: Request, res: Response, next: NextHandler): Promise<void> {
-		const cookies = cookie.parse(req.headers.cookie ?? '');
-		const token = cookies.refresh_token ?? (req.body as AuthGetDiscordRefreshBody | undefined)?.refresh_token;
-
-		if (!token) {
-			return next(unauthorized('missing refresh token'));
-		}
-
-		const response = await discordOAuth2(req, res, next);
-		if (!response) {
-			return;
-		}
-
-		res.cookie('access_token', response.access_token, {
-			expires: new Date(Date.now() + response.expires_in * 1000),
-			sameSite: 'strict',
-			httpOnly: true,
-			domain: this.config.rootDomain.replace(/h?t?t?p?s?:?\/?\/?/, ''),
-			path: '/',
-		});
-
-		return res.end();
-	}
-}
+export { default } from '../../../../v2/auth/discord/refresh/get';
